# docs/verity/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeResponse": {
        "properties": {
          "authorship_verdict": {
            "default": "INDETERMINATE",
            "title": "Authorship Verdict",
            "type": "string"
          },
          "calibrated_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Probability"
          },
          "characters": {
            "title": "Characters",
            "type": "integer"
          },
          "detector_status": {
            "default": "EXPERIMENTAL",
            "title": "Detector Status",
            "type": "string"
          },
          "limitations": {
            "default": [
              "Text statistics do not establish authorship or factual accuracy. No calibrated detector is available."
            ],
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "method": {
            "default": "deterministic-text-statistics-v1",
            "title": "Method",
            "type": "string"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "paragraphs": {
            "title": "Paragraphs",
            "type": "integer"
          },
          "words": {
            "title": "Words",
            "type": "integer"
          }
        },
        "required": [
          "operation_id",
          "characters",
          "words",
          "paragraphs"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
...
      "AzaeronInferenceResult": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "input_sha256": {
            "title": "Input Sha256",
            "type": "string"
          },
          "input_tokens": {
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "model_revision": {
            "title": "Model Revision",
            "type": "string"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "output": {
            "title": "Output",
            "type": "string"
          },
          "output_tokens": {
            "minimum": 0.0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "model_id",
          "model_revision",
          "prompt_version",
          "input_sha256",
          "output",
          "input_tokens",
          "output_tokens",
          "duration_ms"
        ],
        "title": "AzaeronInferenceResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RefineRequest": {
        "additionalProperties": false,
        "properties": {
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "text": {
            "maxLength": 60000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "text"
        ],
        "title": "RefineRequest",
        "type": "object"
      },
      "RefineResponse": {
        "properties": {
          "accepted": {
            "default": false,
            "title": "Accepted",
            "type": "boolean"
          },
          "candidate": {
            "$ref": "#/components/schemas/AzaeronInferenceResult"
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationResult"
          }
        },
        "required": [
          "candidate",
          "verification"
        ],
        "title": "RefineResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationResult": {
        "additionalProperties": false,
        "properties": {
          "deterministic_passed": {
            "title": "Deterministic Passed",
            "type": "boolean"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "outcome": {
            "enum": [
              "VERIFIED",
              "VERIFIED_WITH_WARNINGS",
              "REJECTED",
              "UNAVAILABLE"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "semantic_checked": {
            "default": false,
            "title": "Semantic Checked",
            "type": "boolean"
          },
          "verifier_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verifier Model Id"
          },
          "verifier_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verifier Revision"
          }
        },
        "required": [
          "outcome",
          "deterministic_passed"
        ],
        "title": "VerificationResult",
        "type": "object"
      },
      "VerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "maxLength": 60000,
            "minLength": 1,
            "title": "Candidate",
            "type": "string"
          },
          "locked_spans": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "maxItems": 1000,
            "title": "Locked Spans",
            "type": "array"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "original": {
            "maxLength": 60000,
            "minLength": 1,
            "title": "Original",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "original",
          "candidate"
        ],
        "title": "VerifyRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AzaeronAPIKey": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Privacy-first writing intelligence with reviewable evidence",
    "title": "AZAERON VERITY",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/text/analyze": {
      "post": {
        "operationId": "analyze_api_v1_text_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AzaeronAPIKey": []
          }
        ],
        "summary": "Analyze",
        "tags": [
          "Text"
        ],
        "x-api-key-scopes": [
          "text:analyze"
        ]
      }
    },
    "/api/v1/text/refine": {
      "post": {
        "operationId": "refine_api_v1_text_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AzaeronAPIKey": []
          }
        ],
        "summary": "Refine",
        "tags": [
          "Text"
        ],
        "x-api-key-scopes": [
          "text:refine"
        ]
      }
    },
    "/api/v1/text/verify": {
      "post": {
        "operationId": "verify_api_v1_text_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "AzaeronAPIKey": []
          }
        ],
        "summary": "Verify",
        "tags": [
          "Text"
        ],
        "x-api-key-scopes": [
          "text:verify"
        ]
      }
    },
...
  }
}